fix: スタンダードな見出しにスタイルを統一
