+++ src/app/(standard)/bingotaikai/style.module.css
